// src/app/zdslogic-ui-plugins/shop/core/interfaces/sku.model.ts
export interface Sku {

	skuId: string;
	activeEndDate: string;
	activeStartDate: string;
	addlProductId: string;
	availableFlag: string;
	containerShape: string;
	containerSize: string;
	cost: string;
	currencyCode: string;
	defaultProductId: string;
	depth: string;
	description: string;
	dimensionUnitOfMeasure: string;
	discountableFlag: string;
	displayTemplate: string;
	externalId: string;
	fulfillmentType: string;
	girth: string;
	height: string;
	inventoryType: string;
	isMachineSortable: string;
	longDescription: string;
	name: string;
	quantityAvailable: string;
	retailPrice: string;
	salePrice: string;
	taxCode: string;
	taxableFlag: string;
	upc: string;
	urlKey: string;
	weight: string;
	weightUnitOfMeasure: string;
	width: string;

}